rename generic handler to handler
